requirements.txt
fastapi
uvicorn
pymupdf
openai
faiss-cpu
pydantic
python-dotenv
langchain
tqdm
pinecone-client
